Branch HEAD b=12974 i=nikita i=adilger
1) set as "inline" for such large function ("obd_fail_check()") is not suitable,
   replaced with normal function "__obd_fail_check_set()".
2) atomic operations for obd_fail_count (replace "static count" in obd_fail_check).
3) use "test_and_set_bit()" for "OBD_FAILED/OBD_FAIL_ONCE" check and set to avoid
   unexpected racing condition between multi-thread ops.
4) some cleanup for OBD_FAIL_XXX macro.

+++ lustre/obdclass/darwin/darwin-sysctl.c
@@ -22,7 +22,7 @@ cfs_sysctl_table_header_t *obd_table_header = NULL;
 
 int proc_fail_loc SYSCTL_HANDLER_ARGS;
 int proc_obd_timeout SYSCTL_HANDLER_ARGS;
-extern unsigned int obd_fail_loc;
+extern unsigned long obd_fail_loc;
 extern unsigned int obd_dump_on_timeout;
 extern unsigned int obd_timeout;
 extern unsigned int ldlm_timeout;
@@ -76,7 +76,7 @@ extern cfs_waitq_t obd_race_waitq;
 int proc_fail_loc SYSCTL_HANDLER_ARGS
 { 
 	int error = 0; 
-	int old_fail_loc = obd_fail_loc;
+	long old_fail_loc = obd_fail_loc;
 	
 	error = sysctl_handle_long(oidp, oidp->oid_arg1, oidp->oid_arg2, req); 
 	if (!error && req->newptr != USER_ADDR_NULL) {
